data-deletion

+++ src/admin/AllProducts.jsx
@@ -1,21 +1,36 @@
-import React from "react"
 import { Container, Row, Col } from "reactstrap"
-import { db } from "../firebase.config"
 import { doc, deleteDoc } from "firebase/firestore"
-import useGetData from "../custom-hooks/useGetData"
+import { ref, deleteObject } from 'firebase/storage'
 import { motion } from "framer-motion"
 import { toast } from 'react-toastify'
 
+import { db, storage } from "../firebase.config"
+import useGetData from "../custom-hooks/useGetData"
+
 const AllProducts = () => {
 
-  const {data: productsData, loading} = useGetData("products")
+  const { data: productsData, loading } = useGetData("products")
 
   const deleteProduct = async id => {
 
     const docRef = doc(db, "products", id)
 
-    await deleteDoc(docRef)
-    toast.success("Deleted")
+    for (let i = 0; i <= productsData.length; i++) {
+      if (productsData[i].id === id) {
+
+        const productPath = `productImages/` + productsData[i].productName
+        const Ref = ref(storage, productPath)
+
+        await deleteDoc(docRef)
+        toast.success("Deleted")
+
+        await deleteObject(Ref).then(() => {
+          console.log('success image');
+        }).catch((err) => {
+          console.error('error');
+        })
+      }
+    }
   }
 
   return (
@@ -40,16 +55,16 @@ const AllProducts = () => {
                   ) : (
                     <>
                       {
-                        productsData.map((item)  => (
+                        productsData.map((item) => (
                           <tr key={item.id}>
                             <td><img src={item.imgUrl} alt="" /></td>
-                            <td>{item.title}</td>
+                            <td>{item.productName}</td>
                             <td>{item.category}</td>
                             <td>${item.price}</td>
                             <td>
                               <motion.button
-                                whileTap={{scale: 1.1}} 
-                                onClick={() => {deleteProduct(item.id)}} 
+                                whileTap={{ scale: 1.1 }}
+                                onClick={() => { deleteProduct(item.id) }}
                                 className="btn btn-danger"
                               >
                                 Delete
